requirements.txt
albumentations==1.3.1
fastapi==0.99.1
hydra-core==1.3.2
mlflow==2.4.1
numpy==1.23.5
omegaconf==2.3.0
opencv_python==4.8.0.74
opencv_python_headless==4.8.0.74
starlette==0.28.0
torch==2.0.1
torchmetrics==1.0.0
torchvision==0.15.2
uvicorn==0.22.0
